+++ cdm_reader_mapper/mdf_reader/utils/utilities.py
@@ -85,19 +85,17 @@ def set_missing_values(df: pl.DataFrame) -> pl.DataFrame:
     )
 
 
-def create_mask(df, isna, missing_values=[]):
-    """DOCUMENTATION."""
-    if isna is None:
-        isna = df.isna()
-    valid = df.notna()
-    mask = isna | valid
-    if len(missing_values) > 0:
-        mask[missing_values] = False
-    return mask
-
-
 def adjust_dtype(dtype, df):
     """DOCUMENTATION."""
     if not isinstance(dtype, dict):
         return dtype
     return {k: v for k, v in dtype.items() if k in df.columns}
+
+
+def convert_str_boolean(x):
+    """DOCUMENTATION."""
+    if x == "True":
+        x = True
+    if x == "False":
+        x = False
+    return x